--- a/link-grammar/parse/extract-links.c
+++ b/link-grammar/parse/extract-links.c
@@ -68,6 +68,7 @@ struct extractor_s
 	unsigned int   log2_x_table_size;
 	Pset_bucket ** x_table;  /* Hash table */
 	Parse_set *    parse_set;
+	Word           *words;
 	bool           islands_ok;
 
 	/* thread-safe random number state */
@@ -322,7 +323,7 @@ static Match_node* sort_matchlist(Match_node* mlist)
  * parse structures.
  */
 static
-Parse_set * mk_parse_set(Word* words, fast_matcher_t *mchxt,
+Parse_set * mk_parse_set(fast_matcher_t *mchxt,
                  count_context_t * ctxt,
                  Disjunct *ld, Disjunct *rd, int lw, int rw,
                  Connector *le, Connector *re, unsigned int null_count,
@@ -375,14 +376,14 @@ Parse_set * mk_parse_set(Word* words, fast_matcher_t *mchxt,
 		RECOUNT({xt->set.recount = 0;})
 
 		w = lw + 1;
-		for (int opt = 0; opt <= !!words[w].optional; opt++)
+		for (int opt = 0; opt <= !!pex->words[w].optional; opt++)
 		{
 			null_count += opt;
-			for (dis = words[w].d; dis != NULL; dis = dis->next)
+			for (dis = pex->words[w].d; dis != NULL; dis = dis->next)
 			{
 				if (dis->left == NULL)
 				{
-					pset = mk_parse_set(words, mchxt, ctxt,
+					pset = mk_parse_set(mchxt, ctxt,
 											  dis, NULL, w, rw, dis->right, NULL,
 											  null_count-1, pex);
 					if (pset == NULL) continue;
@@ -393,7 +394,7 @@ Parse_set * mk_parse_set(Word* words, fast_matcher_t *mchxt,
 					RECOUNT({xt->set.recount += pset->recount;})
 				}
 			}
-			pset = mk_parse_set(words, mchxt, ctxt,
+			pset = mk_parse_set(mchxt, ctxt,
 									  NULL, NULL, w, rw, NULL, NULL,
 									  null_count-1, pex);
 			if (pset != NULL)
@@ -459,22 +460,22 @@ Parse_set * mk_parse_set(Word* words, fast_matcher_t *mchxt,
 				for (i=0; i<4; i++) { ls[i] = rs[i] = NULL; }
 				if (Lmatch)
 				{
-					ls[0] = mk_parse_set(words, mchxt, ctxt,
+					ls[0] = mk_parse_set(mchxt, ctxt,
 					             ld, d, lw, w, le->next, d->left->next,
 					             lnull_count, pex);
 
 					if (le->multi)
-						ls[1] = mk_parse_set(words, mchxt, ctxt,
+						ls[1] = mk_parse_set(mchxt, ctxt,
 						              ld, d, lw, w, le, d->left->next,
 						              lnull_count, pex);
 
 					if (d->left->multi)
-						ls[2] = mk_parse_set(words, mchxt, ctxt,
+						ls[2] = mk_parse_set(mchxt, ctxt,
 						              ld, d, lw, w, le->next, d->left,
 						              lnull_count, pex);
 
 					if (le->multi && d->left->multi)
-						ls[3] = mk_parse_set(words, mchxt, ctxt,
+						ls[3] = mk_parse_set(mchxt, ctxt,
 						              ld, d, lw, w, le, d->left,
 						              lnull_count, pex);
 
@@ -485,22 +486,22 @@ Parse_set * mk_parse_set(Word* words, fast_matcher_t *mchxt,
 
 				if (Rmatch && (ls_exists || le == NULL))
 				{
-					rs[0] = mk_parse_set(words, mchxt, ctxt,
+					rs[0] = mk_parse_set(mchxt, ctxt,
 					                 d, rd, w, rw, d->right->next, re->next,
 					                 rnull_count, pex);
 
 					if (d->right->multi)
-						rs[1] = mk_parse_set(words, mchxt, ctxt,
+						rs[1] = mk_parse_set(mchxt, ctxt,
 					                 d, rd, w, rw, d->right, re->next,
 						              rnull_count, pex);
 
 					if (re->multi)
-						rs[2] = mk_parse_set(words, mchxt, ctxt,
+						rs[2] = mk_parse_set(mchxt, ctxt,
 						              d, rd, w, rw, d->right->next, re,
 						              rnull_count, pex);
 
 					if (d->right->multi && re->multi)
-						rs[3] = mk_parse_set(words, mchxt, ctxt,
+						rs[3] = mk_parse_set(mchxt, ctxt,
 						              d, rd, w, rw, d->right, re,
 						              rnull_count, pex);
 				}
@@ -523,7 +524,7 @@ Parse_set * mk_parse_set(Word* words, fast_matcher_t *mchxt,
 				if (ls_exists)
 				{
 					/* Evaluate using the left match, but not the right */
-					Parse_set* rset = mk_parse_set(words, mchxt, ctxt,
+					Parse_set* rset = mk_parse_set(mchxt, ctxt,
 					                        d, rd, w, rw, d->right, re,
 					                        rnull_count, pex);
 					if (rset != NULL)
@@ -545,7 +546,7 @@ Parse_set * mk_parse_set(Word* words, fast_matcher_t *mchxt,
 				     rs[1] != NULL || rs[2] != NULL || rs[3] != NULL))
 				{
 					/* Evaluate using the right match, but not the left */
-					Parse_set* lset = mk_parse_set(words, mchxt, ctxt,
+					Parse_set* lset = mk_parse_set(mchxt, ctxt,
 					                        ld, d, lw, w, le, d->left,
 					                        lnull_count, pex);
 
@@ -625,10 +626,11 @@ bool build_parse_set(extractor_t* pex, Sentence sent,
                     count_context_t *ctxt,
                     unsigned int null_count, Parse_Options opts)
 {
+	pex->words = sent->word;
 	pex->islands_ok = opts->islands_ok;
 
 	pex->parse_set =
-		mk_parse_set(sent->word, mchxt, ctxt,
+		mk_parse_set(mchxt, ctxt,
 		             NULL, NULL, -1, sent->length, NULL, NULL, null_count+1,
 		             pex);