The main change here is:
- mk_parse_set(): Optimize Rmatch processing

In the profiling for fix-long. it speeds up `mk_parse_set()` by about 40% (from 13.9% to 7.9% of the total CPU).
Before this patch (legend at the end [**]):
`     242   1.8%  85.4%     1916  13.9% mk_parse_set`
After it:
`     146   1.2%  89.2%     1002   7.9%  mk_parse_set`


However, since it consists only a small part of total CPU time, the batch benchmarks show only small speedup (for some sentences the speed increase is big [*]):
``` text
 +0.2%    en/corpus-basic.batch
 +1.0%    en/corpus-fixes.batch
 +1.4%    ru/corpus-basic.batch
+26.8%    ../debug/prune/ex/en/bms.batch (In vivo studies...) [*]
 +4.3%    data/en/corpus-fix-long.batch
```

fix-long profiling after this PR (per file, only >0.1% are shown):
```
    6158  49.2%  49.2%     6162  49.2% link-grammar/parse/fast-match.c
    3137  25.0%  74.2%     9948  79.4% link-grammar/parse/count.c
    1223   9.8%  84.0%     1223   9.8% link-grammar/connectors.h
     694   5.5%  89.5%      994   7.9% link-grammar/parse/extract-links.c
     447   3.6%  93.1%     1093   8.7% link-grammar/parse/prune.c
     220   1.8%  94.8%      240   1.9% link-grammar/post-process/post-process.c
     115   0.9%  96.8%      115   0.9% link-grammar/parse/fast-match.h
     109   0.9%  97.7%      266   2.1% link-grammar/parse/preparation.c
      61   0.5%  98.2%       95   0.8% link-grammar/disjunct-utils.c
      43   0.3%  98.9%       58   0.5% link-grammar/string-id.c
      14   0.1%  99.2%       59   0.5% link-grammar/prepare/build-disjuncts.c
      14   0.1%  99.3%       45   0.4% link-grammar/utilities.h
       9   0.1%  99.4%       49   0.4% link-grammar/memory-pool.c
       7   0.1%  99.6%       55   0.4% link-grammar/utilities.c
```

So there is no much room for more speedup for long sentences (only maybe 20%) unless processing shortcuts will be found. The profiling for short sentences is totally different and there is more room there for speedup (>0.5% are shown):
``` text
     227  20.5%  44.5%      264  23.8% link-grammar/post-process/post-process.c
     131  11.8%  56.4%      294  26.6% link-grammar/parse/prune.c
      65   5.9%  62.2%      118  10.7% link-grammar/prepare/exprune.c
      51   4.6%  66.8%       51   4.6% link-grammar/connectors.h
      44   4.0%  70.8%       90   8.1% link-grammar/parse/count.c
      35   3.2%  74.0%      132  11.9% link-grammar/dict-common/dict-utils.c
      29   2.6%  76.6%       29   2.6% link-grammar/parse/fast-match.c
      27   2.4%  79.0%       53   4.8% link-grammar/disjunct-utils.c
      24   2.2%  81.2%      116  10.5% link-grammar/prepare/build-disjuncts.c
      21   1.9%  83.1%      160  14.5% link-grammar/parse/preparation.c
      20   1.8%  84.9%       28   2.5% link-grammar/parse/extract-links.c
      15   1.4%  86.3%       21   1.9% link-grammar/dict-file/read-dict.c
      11   1.0%  89.5%       15   1.4% link-grammar/post-process/pp_linkset.c
      10   0.9%  90.4%       45   4.1% link-grammar/connectors.c
      10   0.9%  91.3%       64   5.8% link-grammar/memory-pool.c
       8   0.7%  93.0%       40   3.6% link-grammar/utilities.h
       5   0.5%  94.5%      158  14.3% link-grammar/tokenize/tokenize.c
```
[**] Fields in the above:

1. Number of profiling samples in this function
2. Percentage of profiling samples in this function
3. Percentage of profiling samples in the functions printed so far
4. Number of profiling samples in this function and its callees
5. Percentage of profiling samples in this function and its callees
6. Name