@linas 2018-11-21
Huh. That was interesting.

When I read through your code changes, its very counter-intuitive. Naively, I figured that passing args on stack would be faster, and since there is a lot of memory-locality on stack, there would be no cache misses. The only overhead is to actually place arguments on stack, and this overhead can be minimized by placing the non-changing once first.  Allowing e.g. tail-recursion optimizations by the compiler.  So your changes suggest that my intuition is dead-wrong.  Interesting...

There are a few other places in the code, where lots of args are passed in the call, instead of being held in a struct. Perhaps those can be sped up, as well. 

Perhaps a really really smart compiler engineer could find some way of automatically doing this kind of optimization...