@ampli 2018-11-22
Maybe you are not wrong. I compared the assembly output (using https://godbolt.org) and when I removed the `island_ok` argument I could not find a difference in the `mk_parse_set()` call sequence. However, the number of total instructions in `mk_parse_set()` got reduced by 7 (I could not get a clear picture about the nature of this reduction with a fast glance at the `diff`, since there are some additions and some removals at different places).
This removal and the removal of "words" gave me a small (but consistent) speedup in 7x7 benchmaks of fix-long. However, I observed a slowdown when In tried to remove the `fast_matcher_t` argument.

Absolutely most of the speedup was from the first commit.