@linas 2018-11-24
Ah. Compilers always have a cutoff: "if fewer than N arguments do this, else do that", together with variants for N occupying different numbers of registers, different number of bytes on stack, etc. If N is too large, everything gets copied to stack; if the argument order is "wrong" then it has to copy just to re-arrange the order into the correct, desired order.  There's a lot of mov, push, pop silliness that comes from that, time-wasting reshuffling of arguments.